services/auth_gateway_service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
pydantic==2.10.3
pydantic-settings==2.6.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
httpx==0.28.1
python-dotenv==1.0.1
cryptography==44.0.0
